# openapi.yaml
{
  "openapi":"3.0.1",
  "info":{
    "title":"OpenAPI definition",
    "version":"v0"
  },
  "servers":[
    {
      "url":"http://localhost:8080",
      "description":"Generated server url"
    }
  ],
  "paths":{
    "/api/transactions/v1/phone-bill-payment/{accountNumber}":{
      "post":{
        "tags":[
          "transaction-resource"
        ],
        "summary":"Perform a phone bill payment transaction",
        "description":"Pays a phone bill using the provided account number, provider name, and bill number.",
        "operationId":"phoneBillPaymentTransaction",
        "parameters":[
          {
            "name":"accountNumber",
            "in":"path",
            "required":true,
            "schema":{
              "type":"string"
            }
          }
        ],
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/BillPaymentTransactionRequest"
              }
            }
          },
          "required":true
        },
        "responses":{
          "404":{
            "description":"Account not found",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/TransactionDto"
                }
              }
            }
          },
          "500":{
            "description":"Internal server error",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/TransactionDto"
                }
              }
            }
          },
          "400":{
            "description":"Invalid input data",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/TransactionDto"
                }
              }
            }
          },
          "200":{
            "description":"Successfully processed phone bill payment",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/TransactionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/transactions/v1/debit/{accountNumber}":{
      "post":{
        "tags":[
          "transaction-resource"
        ],
        "summary":"Perform a debit transaction",
        "description":"Debits a specified amount from the given account number.",
        "operationId":"debitTransaction",
        "parameters":[
          {
            "name":"accountNumber",
            "in":"path",
            "required":true,
            "schema":{
              "type":"string"
            }
          }
        ],
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/TransactionRequest"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"Successfully debited the account",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/TransactionDto"
                }
              }
            }
          },
          "404":{
            "description":"Account not found",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/TransactionDto"
                }
              }
            }
          },
          "400":{
            "description":"Invalid input data or insufficient funds",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/TransactionDto"
                }
              }
            }
          },
          "500":{
            "description":"Internal server error",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/TransactionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/transactions/v1/credit/{accountNumber}":{
      "post":{
        "tags":[
          "transaction-resource"
        ],
        "summary":"Perform a credit transaction",
        "description":"Credits a specified amount to the given account number.",
        "operationId":"creditTransaction",
        "parameters":[
          {
            "name":"accountNumber",
            "in":"path",
            "required":true,
            "schema":{
              "type":"string"
            }
          }
        ],
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/TransactionRequest"
              }
            }
          },
          "required":true
        },
        "responses":{
          "404":{
            "description":"Account not found",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/TransactionDto"
                }
              }
            }
          },
          "500":{
            "description":"Internal server error",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/TransactionDto"
                }
              }
            }
          },
          "400":{
            "description":"Invalid input data",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/TransactionDto"
                }
              }
            }
          },
          "200":{
            "description":"Successfully credited the account",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/TransactionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/bank-accounts/v1/create":{
      "post":{
        "tags":[
          "bank-account-resource"
        ],
        "summary":"Create a new bank account",
        "description":"Creates a new bank account for the given owner and mail. Returns the details of the created bank account.",
        "operationId":"createBankAccount",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/BankAccountRequest"
              }
            }
          },
          "required":true
        },
        "responses":{
          "500":{
            "description":"Internal server error",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/BankAccountDto"
                }
              }
            }
          },
          "400":{
            "description":"Invalid input data",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/BankAccountDto"
                }
              }
            }
          },
          "200":{
            "description":"Successfully created the bank account",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/BankAccountDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/bank-accounts/v1/{bankAccountNumber}":{
      "get":{
        "tags":[
          "bank-account-resource"
        ],
        "summary":"Retrieve a bank account",
        "description":"Fetches the details of a bank account using the provided bank account number.",
        "operationId":"getBankAccount",
        "parameters":[
          {
            "name":"bankAccountNumber",
            "in":"path",
            "required":true,
            "schema":{
              "type":"string"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"Successfully retrieved the bank account details",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/BankAccountDto"
                }
              }
            }
          },
          "404":{
            "description":"Bank account not found",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/BankAccountDto"
                }
              }
            }
          },
          "500":{
            "description":"Internal server error",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/BankAccountDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components":{
    "schemas":{
      "BillPaymentTransactionRequest":{
        "type":"object",
        "properties":{
          "amount":{
            "type":"number",
            "format":"double"
          },
          "providerName":{
            "type":"string"
          },
          "billNumber":{
            "type":"string"
          }
        }
      },
      "TransactionDto":{
        "type":"object",
        "properties":{
          "date":{
            "type":"string",
            "format":"date-time"
          },
          "amount":{
            "type":"number",
            "format":"double"
          },
          "type":{
            "type":"string",
            "enum":[
              "DEPOSIT_TRANSACTION",
              "WITHDRAWAL_TRANSACTION",
              "PHONE_BILL_PAYMENT_TRANSACTION"
            ]
          },
          "approvalCode":{
            "type":"string"
          }
        }
      },
      "TransactionRequest":{
        "type":"object",
        "properties":{
          "amount":{
            "type":"number",
            "format":"double"
          }
        }
      },
      "BankAccountRequest":{
        "type":"object",
        "properties":{
          "owner":{
            "type":"string"
          },
          "mail":{
            "type":"string"
          }
        }
      },
      "BankAccountDto":{
        "type":"object",
        "properties":{
          "owner":{
            "type":"string"
          },
          "accountNumber":{
            "type":"string"
          },
          "balance":{
            "type":"number",
            "format":"double"
          },
          "transactionDtoList":{
            "type":"array",
            "items":{
              "$ref":"#/components/schemas/TransactionDto"
            }
          },
          "mail":{
            "type":"string"
          }
        }
      }
    }
  }
}